web: Split elements into separate files.
