+++ web/src/elements/Interface/index.ts
@@ -1,4 +1,5 @@
-import { AuthenticatedInterface, Interface } from "./Interface";
+import { AuthenticatedInterface } from "./AuthenticatedInterfaceElement.js";
+import { Interface } from "./InterfaceElement.js";
 
 export { Interface, AuthenticatedInterface };
 export default Interface;
